fix use of applicationContext.getBeansOfType instead of the old lookupMap. The key is now interface#hint whereas was only hint with plexus
git-svn-id: https://svn.apache.org/repos/asf/archiva/trunk@1130206 13f79535-47bb-0310-9956-ffa450edef68

+++ archiva-modules/archiva-base/archiva-proxy/src/test/java/org/apache/maven/archiva/proxy/AbstractProxyTestCase.java
@@ -157,7 +157,10 @@ public abstract class AbstractProxyTestCase
 
         // Setup source repository (using legacy layout)
         repoLocation = new File( REPOPATH_LEGACY_MANAGED_TARGET );
-        FileUtils.deleteDirectory( repoLocation );
+        if (repoLocation.exists())
+        {
+            FileUtils.deleteDirectory( repoLocation );
+        }
         copyDirectoryStructure( new File( REPOPATH_LEGACY_MANAGED ), repoLocation );
 
         managedLegacyRepository =
